+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   * flutter pub upgrade --major-versions
   * Created TH2FileEditElementEditController.optionTypeBeingEdited.
   * Add keys to radio groups at MPPLATypeOptionsOverlayWindowWidget.
+  * Created 3500-ui_change_line_type_test.
 
 ## 0.2.35 - 2026-02-13 - The [Chiquita Bacana's daughter](https://www.youtube.com/watch?v=QajgvvuFOSg) release
 * New features: